LayoutTests/fast/shadow-dom/fullscreen-in-shadow-fullscreenElement-expected.txt

Test that webkitFullscreenElement retargets the fullscreen element inside a shadow tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.webkitFullscreenElement is shadowHost
PASS successfullyParsed is true

TEST COMPLETE
Enter fullscreen
